SBERDOMA-62 user.User move gql
