+++ apps/condo/utils/serverSchema/User.js
@@ -1,11 +1,12 @@
+const { User: UserGQL } = require('@condo/domains/user/gql')
+
 const { execGqlWithoutAccess } = require('./utils')
-const { User } = require('../../gql/User')
 
 async function findUser (context, query) {
     if (!context) throw new Error('no context')
     if (!query) throw new Error('no query')
     return await execGqlWithoutAccess(context, {
-        query: User.GET_ALL_OBJS_QUERY,
+        query: UserGQL.GET_ALL_OBJS_QUERY,
         variables: {
             where: query,
         },
